🧹 [Code Health] Refactor Sidebar component
